refactor: split StructureEditor into AuthPanel and TreeRow sub-components, remove duplicate handler
