+++ src/components/StructureEditor/index.jsx
@@ -7,13 +7,12 @@ import {
   createStructurePR,
   fetchRawFile,
   verifyGitHubToken,
-  loginWithGitHub,
-  startDeviceFlow,
-  pollDeviceFlow,
 } from '@site/src/utils/github';
 import mammoth from 'mammoth';
 import { WysiwygEditor, mdToHtml, htmlToMd } from '@site/src/components/EditModal';
 import { splitFrontmatter, slugify, setFrontmatterField } from '@site/src/utils/markdown';
+import { AuthPanel } from './AuthPanel';
+import { InlineForm, TreeRow } from './TreeRow';
 import styles from './index.module.css';
 
 const AUTH_KEY = 'masakhane_pb_auth';
@@ -62,384 +61,6 @@ function starterCategory(label, position, description = '') {
     { label, position, link: { type: 'generated-index', description } },
     null,
     2,
-  );
-}
-
-/* ── Inline add / rename form ────────────────────────────────────────── */
-
-function InlineForm({ placeholder, initialValue = '', onConfirm, onCancel }) {
-  const [value, setValue] = useState(initialValue);
-  const inputRef = useRef(null);
-
-  useEffect(() => {
-    inputRef.current?.focus();
-    inputRef.current?.select();
-  }, []);
-
-  function handleKey(e) {
-    if (e.key === 'Enter') { e.preventDefault(); if (value.trim()) onConfirm(value.trim()); }
-    if (e.key === 'Escape') onCancel();
-  }
-
-  return (
-    <div className={styles.inlineForm}>
-      <input
-        ref={inputRef}
-        className={styles.inlineInput}
-        value={value}
-        onChange={e => setValue(e.target.value)}
-        onKeyDown={handleKey}
-        placeholder={placeholder}
-      />
-      <button
-        className={styles.inlineConfirm}
-        onClick={() => value.trim() && onConfirm(value.trim())}
-        type="button"
-        disabled={!value.trim()}
-      >
-        Add
-      </button>
-      <button className={styles.inlineCancel} onClick={onCancel} type="button">✕</button>
-    </div>
-  );
-}
-
-/* ── GitHub auth panel ───────────────────────────────────────────────── */
-
-const GH_MARK = (
-  <svg width="20" height="20" viewBox="0 0 16 16" fill="currentColor" aria-hidden="true" style={{ flexShrink: 0 }}>
-    <path d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.013 8.013 0 0016 8c0-4.42-3.58-8-8-8z"/>
-  </svg>
-);
-
-function AuthPanel({ auth, clientId, proxyUrl, callbackUrl, onConnect, onDisconnect }) {
-  // phases: 'idle' | 'loading' | 'device-pending' | 'waiting'
-  const [phase, setPhase] = useState('idle');
-  const [deviceInfo, setDeviceInfo] = useState(null);
-  const [error, setError] = useState('');
-  const [draft, setDraft] = useState('');
-  const abortRef = useRef(null);
-
-  // What's available depends on which env vars are set
-  const canPopup  = !!(clientId && proxyUrl && callbackUrl);
-  const canDevice = !!(clientId && proxyUrl);
-
-  // ── Popup OAuth (best UX — requires Cloudflare Worker + callbackUrl) ──
-  async function handlePopupOAuth() {
-    setPhase('loading');
-    setError('');
-    try {
-      const token = await loginWithGitHub(clientId, proxyUrl, callbackUrl);
-      await onConnect(token);
-    } catch (e) {
-      if (e.message === 'cancelled') { setPhase('idle'); return; }
-      // Popup was blocked → fall back to device flow silently
-      if (e.message?.includes('Popup')) {
-        setError('Popup blocked — switching to device flow.');
-        await handleDeviceFlow();
-        return;
-      }
-      setError(e.message || 'GitHub login failed.');
-      setPhase('idle');
-    }
-  }
-
-  // ── Device Flow (requires Cloudflare Worker, no client_secret) ────────
-  async function handleDeviceFlow() {
-    setPhase('loading');
-    setError('');
-    try {
-      const info = await startDeviceFlow(clientId, proxyUrl);
-      setDeviceInfo(info);
-      window.open(info.verification_uri, '_blank', 'noopener,noreferrer');
-      setPhase('device-pending');
-      const ctrl = new AbortController();
-      abortRef.current = ctrl;
-      const token = await pollDeviceFlow(clientId, info.device_code, proxyUrl, info.interval || 5, ctrl.signal);
-      await onConnect(token);
-    } catch (e) {
-      if (e.message === 'cancelled') { setPhase('idle'); return; }
-      setError(e.message || 'GitHub login failed.');
-      setPhase('idle');
-    }
-  }
-
-  function cancelDevice() {
-    abortRef.current?.abort();
-    setPhase('idle');
-    setDeviceInfo(null);
-    setError('');
-  }
-
-  // ── Main sign-in handler: picks best available method ─────────────────
-  function handleSignIn() {
-    if (canPopup)  return handlePopupOAuth();
-    if (canDevice) return handleDeviceFlow();
-  }
-
-  async function handlePaste() {
-    if (!draft.trim()) return;
-    setPhase('loading');
-    setError('');
-    try {
-      await onConnect(draft.trim());
-    } catch (e) {
-      setError(e.message || 'Invalid token or missing public_repo permission.');
-      setPhase('idle');
-    }
-  }
-
-  // ── Connected ─────────────────────────────────────────────────────────
-  if (auth) {
-    return (
-      <div className={styles.authConnected}>
-        <img src={auth.avatarUrl} alt={auth.username} className={styles.authAvatar} />
-        <span className={styles.authUsername}>@{auth.username}</span>
-        <button className={styles.authSignOutBtn} onClick={onDisconnect} type="button">Sign out</button>
-      </div>
-    );
-  }
-
-  // ── Device flow pending ───────────────────────────────────────────────
-  if (phase === 'device-pending' && deviceInfo) {
-    return (
-      <div className={styles.authDeviceBox}>
-        <p className={styles.authDevicePrompt}>Enter this code on the GitHub tab that just opened:</p>
-        <div className={styles.authDeviceCode}>{deviceInfo.user_code}</div>
-        <p className={styles.authDeviceHint}>Waiting for authorisation…</p>
-        <button className={styles.authCancelSmall} onClick={cancelDevice} type="button">Cancel</button>
-      </div>
-    );
-  }
-
-  // ── Default: GitHub button + token input always visible ──────────────
-  return (
-    <div className={styles.authBlock}>
-      <button
-        className={styles.authGitHubBtn}
-        onClick={handleSignIn}
-        disabled={phase === 'loading' || !canDevice}
-        title={!canDevice ? 'OAuth not configured — use a personal access token below' : undefined}
-        type="button"
-      >
-        {GH_MARK}
-        {phase === 'loading' ? 'Connecting…' : 'Sign in with GitHub'}
-      </button>
-      <p className={styles.authOrDivider}>or use a personal access token</p>
-      <div className={styles.authPatRow}>
-        <input
-          type="password"
-          className={styles.authTokenInput}
-          value={draft}
-          onChange={e => setDraft(e.target.value)}
-          onKeyDown={e => e.key === 'Enter' && handlePaste()}
-          placeholder="ghp_xxxxxxxxxxxx"
-          autoComplete="off"
-        />
-        <button
-          className={styles.authConnectBtn}
-          onClick={handlePaste}
-          disabled={!draft.trim() || phase === 'loading'}
-          type="button"
-        >
-          {phase === 'loading' ? '…' : 'Connect'}
-        </button>
-      </div>
-      <a
-        href="https://github.com/settings/tokens/new?scopes=public_repo&description=Masakhane+Playbook"
-        target="_blank" rel="noreferrer noopener"
-        className={styles.authTokenLink}
-      >
-        Generate a token on GitHub ↗
-      </a>
-      {error && <p className={styles.authError}>{error}</p>}
-    </div>
-  );
-}
-
-/* ── Single tree row ─────────────────────────────────────────────────── */
-
-function TreeRow({
-  node,
-  depth,
-  siblings,
-  expanded,
-  onToggle,
-  activeForm,
-  onSetActiveForm,
-  onAddPage,
-  onAddSection,
-  onAddSubPage,
-  onRename,
-  onMoveUp,
-  onMoveDown,
-  onDelete,
-  onEditPage,
-  onEditSection,
-  editingPath,
-  loadingPaths,
-  locked,
-}) {
-  const idx = siblings.indexOf(node);
-  const canUp = idx > 0;
-  const canDown = idx < siblings.length - 1;
-  const isExpanded = expanded.has(node.path);
-  const isLoading = loadingPaths.has(node.path);
-  const formKey = `${node.path}:`;
-  const isEditing = editingPath === node.path;
-  const lockTitle = 'Sign in with GitHub to make changes';
-
-  const indent = depth * 20;
-
-  return (
-    <>
-      <div
-        className={`${styles.treeRow} ${isEditing ? styles.treeRowActive : ''}`}
-        style={{ paddingLeft: indent + 8 }}
-      >
-        {node.type === 'section' ? (
-          <button className={styles.expandBtn} onClick={() => onToggle(node.path)} type="button" aria-label={isExpanded ? 'Collapse' : 'Expand'}>
-            {isExpanded ? '▾' : '▸'}
-          </button>
-        ) : (
-          <span className={styles.expandSpacer} />
-        )}
-
-        <span className={styles.nodeIcon}>{node.type === 'section' ? '📁' : '📄'}</span>
-        <span
-          className={`${styles.nodeLabel} ${styles.nodeLabelClickable}`}
-          title={node.path}
-          onClick={() => node.type === 'page' ? onEditPage(node) : onEditSection(node)}
-        >
-          {node.label}
-        </span>
-
-        {isLoading && <span className={styles.loadingDot} title="Loading…">⋯</span>}
-
-        <div className={styles.nodeActions}>
-          {node.type === 'page' && (
-            <button
-              className={`${styles.actionBtn} ${styles.editPageBtn}`}
-              title={locked ? lockTitle : 'Edit page content'}
-              onClick={locked ? undefined : () => onEditPage(node)}
-              disabled={locked}
-              type="button"
-            >
-              ✎
-            </button>
-          )}
-          {node.type === 'section' && (
-            <button
-              className={`${styles.actionBtn} ${styles.editPageBtn}`}
-              title={locked ? lockTitle : 'Edit section intro page'}
-              onClick={locked ? undefined : () => onEditSection(node)}
-              disabled={locked}
-              type="button"
-            >
-              ✎
-            </button>
-          )}
-          {node.type === 'section' && (
-            <button
-              className={styles.actionBtn}
-              title={locked ? lockTitle : 'Add page inside this section'}
-              onClick={locked ? undefined : () => onSetActiveForm(`${formKey}add-page`)}
-              disabled={locked}
-              type="button"
-            >
-              + Page
-            </button>
-          )}
-          {node.type === 'page' && (
-            <button
-              className={styles.actionBtn}
-              title={locked ? lockTitle : 'Add a sub-page'}
-              onClick={locked ? undefined : () => onSetActiveForm(`${formKey}add-subpage`)}
-              disabled={locked}
-              type="button"
-            >
-              + Sub
-            </button>
-          )}
-          <button
-            className={styles.actionBtn}
-            title={locked ? lockTitle : 'Rename'}
-            onClick={locked ? undefined : () => onSetActiveForm(`${formKey}rename`)}
-            disabled={locked}
-            type="button"
-          >
-            ✏
-          </button>
-          <button className={`${styles.actionBtn} ${styles.arrowBtn}`} title={locked ? lockTitle : 'Move up'} onClick={locked ? undefined : () => onMoveUp(node)} disabled={locked || !canUp} type="button">↑</button>
-          <button className={`${styles.actionBtn} ${styles.arrowBtn}`} title={locked ? lockTitle : 'Move down'} onClick={locked ? undefined : () => onMoveDown(node)} disabled={locked || !canDown} type="button">↓</button>
-          {node.type === 'page' && (
-            <button
-              className={`${styles.actionBtn} ${styles.deleteBtn}`}
-              title={locked ? lockTitle : 'Delete page'}
-              onClick={locked ? undefined : () => onDelete(node)}
-              disabled={locked}
-              type="button"
-            >
-              🗑
-            </button>
-          )}
-        </div>
-      </div>
-
-      {activeForm === `${formKey}add-page` && (
-        <div style={{ paddingLeft: indent + 36 }}>
-          <InlineForm
-            placeholder="New page title…"
-            onConfirm={label => { onAddPage(node, label); onSetActiveForm(null); }}
-            onCancel={() => onSetActiveForm(null)}
-          />
-        </div>
-      )}
-      {activeForm === `${formKey}add-subpage` && (
-        <div style={{ paddingLeft: indent + 36 }}>
-          <InlineForm
-            placeholder="New sub-page title…"
-            onConfirm={label => { onAddSubPage(node, label); onSetActiveForm(null); }}
-            onCancel={() => onSetActiveForm(null)}
-          />
-        </div>
-      )}
-      {activeForm === `${formKey}rename` && (
-        <div style={{ paddingLeft: indent + 36 }}>
-          <InlineForm
-            placeholder="New label…"
-            initialValue={node.label}
-            onConfirm={label => { onRename(node, label); onSetActiveForm(null); }}
-            onCancel={() => onSetActiveForm(null)}
-          />
-        </div>
-      )}
-
-      {node.type === 'section' && isExpanded && node.children.map(child => (
-        <TreeRow
-          key={child.path}
-          node={child}
-          depth={depth + 1}
-          siblings={node.children}
-          expanded={expanded}
-          onToggle={onToggle}
-          activeForm={activeForm}
-          onSetActiveForm={onSetActiveForm}
-          onAddPage={onAddPage}
-          onAddSection={onAddSection}
-          onAddSubPage={onAddSubPage}
-          onRename={onRename}
-          onMoveUp={onMoveUp}
-          onMoveDown={onMoveDown}
-          onDelete={onDelete}
-          onEditPage={onEditPage}
-          onEditSection={onEditSection}
-          editingPath={editingPath}
-          loadingPaths={loadingPaths}
-          locked={locked}
-        />
-      ))}
-    </>
   );
 }
 
@@ -505,26 +126,6 @@ export function StructureEditorContent({ onClose }) {
   const [uploadPlacement, setUploadPlacement] = useState('page'); // 'page' | 'section' | 'subsection'
   const [uploadTargetSection, setUploadTargetSection] = useState('');
   const [uploadSectionName, setUploadSectionName] = useState('');
-
-  function handleSplitterMouseDown(e) {
-    e.preventDefault();
-    const startX = e.clientX;
-    const startWidth = leftWidth;
-    document.body.style.cursor = 'col-resize';
-    document.body.style.userSelect = 'none';
-    function onMove(ev) {
-      const next = Math.max(180, Math.min(560, startWidth + ev.clientX - startX));
-      setLeftWidth(next);
-    }
-    function onUp() {
-      document.body.style.cursor = '';
-      document.body.style.userSelect = '';
-      document.removeEventListener('mousemove', onMove);
-      document.removeEventListener('mouseup', onUp);
-    }
-    document.addEventListener('mousemove', onMove);
-    document.addEventListener('mouseup', onUp);
-  }
 
   function handleHeaderMouseDown(e) {
     if (fullscreen) return;
